Move TestSupportedJujuSeries to non linux test file.  Add non ubuntu contents for results.
